testController: Do not parse for testenv on its own but search for the listed testenv
Instead of parsing the tox.ini file for testenv with a regex,
testController uses the output from tox and tries to locate the listed
testenvs (with the regex) within the file to add the test task marker.
Additionally, it adds all testenvs that could not be located (i.e. the
ones from permutations) to line zero. This lets them at least appear
correctly in the tree view.